Wall pivots updated

+++ Navi Admin/Assets/Scripts/WallDotController.cs
@@ -4,8 +4,6 @@ using UnityEngine;
 
 public class WallDotController : MonoBehaviour
 {
-    private const int MAX_LINES = 8;
-
     public List<WallDotController> neighborsDots = new List<WallDotController>();
     public List<GameObject> lines = new List<GameObject>();
     public List<int> linesType = new List<int>();
@@ -15,6 +13,7 @@ public class WallDotController : MonoBehaviour
     private void Update()
     {
         if (lines.Count == 0) DeleteDot();
+        position = this.transform.position;
     }
 
     public void SetPosition(Vector3 _position)
@@ -29,18 +28,10 @@ public class WallDotController : MonoBehaviour
 
     public void AddLine(GameObject _line, int _type, WallDotController _neighborDot)
     {
-        if (linesCount >= MAX_LINES)
-        {
-            print("Max lines reached");
-            return;
-        }
-        else
-        {
-            lines.Add(_line);
-            linesType.Add(_type);
-            neighborsDots.Add(_neighborDot);
-            linesCount++;
-        }
+        lines.Add(_line);
+        linesType.Add(_type);
+        neighborsDots.Add(_neighborDot);
+        linesCount++;
     }
 
     public void DeleteLine(int _index)
@@ -51,12 +42,15 @@ public class WallDotController : MonoBehaviour
         linesCount--;
     }
 
-    public void DeleteDot()
+    public void DeleteDot(bool _deleteLines = true)
     {
-        for (int i = 0; i < linesCount; i++)
+        if (_deleteLines)
         {
-            neighborsDots[i].DeleteLine(neighborsDots[i].neighborsDots.IndexOf(this));
-            Destroy(lines[i]);
+            for (int i = 0; i < linesCount; i++)
+            {
+                neighborsDots[i].DeleteLine(neighborsDots[i].neighborsDots.IndexOf(this));
+                Destroy(lines[i]);
+            }
         }
         Destroy(gameObject);
     }
